Solve edge cases
If pw entered b4 un/email, show err msg only if invalid. If un/email is then entered, err msg for this field will display based on validity & whether it's recognized, but pw err msg will go blank unless it's invalid as un/email is changed. If un/email already entered, then err msg for pw displays if it doesn't match an account or if it's not valid.