Store some userinfo in sessionStorage
These values are set independently from the currentUser object. If user logs in, they are set to be the same as those in existing currentUser. This is done so that changes on edit-user-info form can be compared w/ existing data & b/c input-handler func for those fields set these ind. values, not those vals in currentUser. Also, make small changes so edit-user-info form renders properly (not disappearing on certain logic used on signup form).